오류로 인한 초기화 및 서버 재작업

+++ models/etfstocklist.ts
@@ -53,7 +53,7 @@ export class etfStockList extends Model<etfStockListAttributes, etfStockListCrea
       field: 'reg_date'
     },
     etfPercent: {
-      type: DataTypes.FLOAT(2,2),
+      type: DataTypes.FLOAT,
       allowNull: true,
       comment: "ETF 비율",
       field: 'etf_percent'
